types location refactored
