--- src/lib/types/journal.ts
@@ -1,11 +1,0 @@
-export type Mood = "great" | "good" | "okay" | "bad" | "terrible";
-
-export interface Entry {
-    id: string;
-    mood: Mood;
-    title: string;
-    content: string;
-    createdAt: string;
-    favorited: boolean;
-    tags?: string[];
-}
